rename db to db_file
