+++ testing/test_dbms/sample_dbms/wiggle_number_sample_dbms.txt
@@ -1,1 +1,1 @@
-10
+0
